em64t support
(Logical change 1.141)


git-svn-id: http://svn.fedorahosted.org/svn/corosync/trunk@505 fd59a12c-fef9-0310-b244-a6a79926bd2f

+++ include/evs.h
@@ -36,7 +36,7 @@
 
 #include <netinet/in.h>
 
-typedef unsigned long long evs_handle_t;
+typedef uint64_t evs_handle_t;
 
 typedef enum {
 	EVS_DISPATCH_ONE,
