+++ test/publish.c
@@ -205,7 +205,7 @@ test_pub()
 		exit(result);
 	}
 
-	printf("Published event ID: 0x%llx\n", event_id);
+	printf("Published event ID: 0x%llx\n", (unsigned long long)event_id);
 	}
 
 	/*
@@ -287,7 +287,7 @@ event_callback( SaEvtSubscriptionIdT subscription_id,
 
 	printf("event_callback called\n");
 	printf("sub ID: %x\n", subscription_id);
-	printf("event_handle %llx\n", event_handle);
+	printf("event_handle %llx\n", (unsigned long long)event_handle);
 	printf("event data size %d\n", event_data_size);
 
 	evt_pat_get_array.patternsNumber = 4;
@@ -311,9 +311,9 @@ event_callback( SaEvtSubscriptionIdT subscription_id,
 	}
 
 	printf("priority: 0x%x\n", priority);
-	printf("retention: 0x%llx\n", retention_time);
+	printf("retention: 0x%llx\n", (unsigned long long)retention_time);
 	printf("publisher name content: \"%s\"\n", publisher_name.value); 
-	printf("event id: 0x%llx\n", event_id);
+	printf("event id: 0x%llx\n", (unsigned long long)event_id);
 evt_free:
 	result = saEvtEventFree(event_handle);
 	get_sa_error(result, result_buf, result_buf_len);
